requirements.txt
numpy
imdbpy
pandas